Support aarch64 targets by passing arm64 to npm
In bb files for iotivity-node, smarthome-gateway and
iot-rest-api-server, do_compile() includes a case statement to translate
compiler architecture designations to the strings used by the npm
command for the same purpose (--arch flag). The target architecture for
Dragonboard (and several other 64 bit ARM boards) is aarch64.  aarch64
was not included in the known list, resulting in the printout "unknown
architecture" and stopped operation.

Add "aarch64" and translate it to what npm expects, which is "arm64".

All 3 .bb files changed identically.  Some refactoring might be advised
since the affected code is repetitive.

[GDP-571] Iotivity build failure on Dragonboard

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>